Lowers path priority for `chruby`:
With regard to issue #48, if `chruby` were to have a lower $PATH
priority, user specific "bin" directories such as $HOME/bin would
continue to take precedence should there be a conflict between binaries
in $HOME/bin and ~/.gem/$ruby/$version/bin.

NOTE: this can be updated to such that then precedence change is optional
if that would be more suitable.