Source chruby.sh in the subshell
Bash doesn't export functions to subshells.  This means that the
`source` call to load chruby.sh in chruby-exec wasn't actually doing
anything.

This patch makes chruby-exec source chruby.sh as the last action before
calling the chruby function in the subshell it launches, which
guarantees that it's available even if the user hasn't sourced chruby.sh
in their .bashrc/.zshrc file.