remove superfluous quotes. optimize find calls.
A fair bit of the quoting used in thie script was unnecessary, so
I've taken the liberty of clearing that up to reduce line-noise.

The exiting method to populate the RUBIES array used some unnecessary
forks, so I've switched to a process substitution and a single call to
`find'. This approach still cannot handle paths with embedded newlines
but that edgecase requires GNU find's -print0 or more ugly forks with
-exec printf '%s\0' which defeats the purpose. If GNU find is expected
then altering this to be completely safe is a matter of add -d'' to
the read loop and -print0 to the find args. I've also used an array
to store the args to `find' in order to reduce the line length
considerably. IMO, a better approach would be to use 2-space tabs
instead to reduce the horizontal screen space wasted by excessive
indentation.

As per best practices guidelines, I've also changed the RUBIES var
to `rubies' instead. All-caps variable names are conventionally
reserved for environment variables and shell builtin variables. This
should be considered for all variables which do not need to be
exported.