Update variables to adhere to wider naming conventions
Historically, upper-case variable names were used by shells for
their internal variables (PS1, HISTSIZE, RANDOM, BASH_VERSION)
and for environment variables which need to be passed on to new
commands (EDITOR, VISUAL, PATH, etc). User-defined variables
which are meant to be exclusive to the script should always
include at least 1(one) small-case letter in the identifier
in order to ensure that there are no clashes with existing
environment variables or shell variables.

I have left the array named RUBIES alone thus far, because
arrays cannot be exported and therefore cannot clash with
things inherited from the environment. I would suggest that
this be small-cased instead, but it is practically safe
as-is and would require notifying the existing userbase of
the change. This could result in a headache for no explicit
gain.