Rehash whenever Ruby changes.
If changing to a Ruby that has different gems with binaries, the shell
does not always recognize those changes, e.g., when changing rubies on
shell initialization:

  % pry
  zsh: correct 'pry' to 'pr' [nyae]? n

chruby_reset() does the right thing by rehashing. If there is no value
in $RUBY_ROOT, chruby_use() does not call chruby_reset(). This forces
hash -r to be called if there is no value.