keep binary compatible gem dir (for ruby 2.1)
ruby 2.1.* are binary compatible - therefore it is not needed to recompile
gems when migrating between these versions. So the gem dir should not be
different for these versions.

This patches adjusts the RUBY_VERSION calculation to be 2.1.0 for all 2.1.*
ruby versions. This results in constant gem dirs and the same behavior as
for rubygems and the GEM_ROOT.

To minimize to risk to break something the adjusted paths are protected by a
rescue block and were reexported so that it is ensured that we have an valid
RUBY_VERSION value afterwards.

Because the RbConfig::CONFIG['ruby_version'] declared 1.8.* and 1.9.* as 1.8
or 1.9.1 this value is not used for these rubies.