zsh: also switch rubies after changing pwd
The existing preexec hook switches rubies before
any command. In a new shell, this is good setup.

However, it means that when changing from
directory a to b, we switch based on the
.ruby-version in the folder we're leaving,
not the one we're entering.

Adding a chpwd hook switches again when we
arrive in the new folder, so that we end
up with the right one.