Fix error in jruby
chruby would fail in jruby with a project containing a `.jrubyrc` file
with the `cli.load.gemfile=true` configuration. It would spit
`LoadError: no such file to load -- bundler/setup` and fail to properly
set various environment variables.

I thought of using a `--no-rc` flag, but it does not seem to be provided
by `ruby` binaries. So I figured switching to a non-project directory
could work.